apps/worker/requirements.txt
rq==1.15.1
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2
PyYAML==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0